+++ client/my-sites/exporter/index.jsx
@@ -1,62 +1,40 @@
 /**
  * External dependencies
  */
+import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import flowRight from 'lodash/flowRight';
 
 /**
  * Internal dependencies
  */
-import analytics from 'lib/analytics';
 import config from 'config';
-import Exporter from './exporter';
-import { getSiteSlug } from 'state/sites/selectors';
-import {
-	shouldShowProgress,
-	getSelectedPostType,
-	isExporting,
-	getExportingState,
-} from 'state/site-settings/exporter/selectors';
-import {
-	advancedSettingsFetch,
-	exportStatusFetch,
-	setPostType,
-	startExport,
-} from 'state/site-settings/exporter/actions';
-import { States } from 'state/site-settings/exporter/constants';
+import Notices from './notices';
+import ExportCard from './export-card';
+import GuidedTransferCard from './guided-transfer-card';
+import InProgressCard from './guided-transfer-card/in-progress';
 
-function mapStateToProps( state ) {
-	const siteId = state.ui.selectedSiteId;
+class Exporter extends Component {
+	render() {
+		const { isGuidedTransferInProgress } = this.props;
+		const showGuidedTransferOptions = config.isEnabled( 'manage/export/guided-transfer' );
 
+		return <div className="exporter">
+			<Notices />
+			{ showGuidedTransferOptions && isGuidedTransferInProgress &&
+				<InProgressCard /> }
+			<ExportCard />
+			{ showGuidedTransferOptions && ! isGuidedTransferInProgress &&
+				<GuidedTransferCard /> }
+		</div>;
+	}
+}
+
+function mapStateToProps() {
 	return {
-		siteId,
-		siteSlug: getSiteSlug( state, siteId ),
-		postType: getSelectedPostType( state ),
-		shouldShowProgress: shouldShowProgress( state, siteId ),
-		isExporting: isExporting( state, siteId ),
-		downloadURL: state.siteSettings.exporter.downloadURL,
-		didComplete: getExportingState( state, siteId ) === States.COMPLETE,
-		didFail: getExportingState( state, siteId ) === States.FAILED,
-		showGuidedTransferOptions: config.isEnabled( 'manage/export/guided-transfer' ),
-
 		// This will be replaced with a Redux selector once we've built out
 		// the reducers
-		isGuidedTransferInProgress: false,
+		isGuidedTransferInProgress: true,
 	};
 }
 
-function mapDispatchToProps( dispatch ) {
-	return {
-		advancedSettingsFetch: flowRight( dispatch, advancedSettingsFetch ),
-		exportStatusFetch: flowRight( dispatch, exportStatusFetch ),
-		setPostType: flowRight( dispatch, setPostType ),
-		startExport: ( siteId, options ) => {
-			analytics.tracks.recordEvent( 'calypso_export_start_button_click', {
-				scope: ( options && options.exportAll === false ) ? 'selected' : 'all'
-			} );
-			dispatch( startExport( siteId, options ) );
-		}
-	};
-}
-
-export default connect( mapStateToProps, mapDispatchToProps )( Exporter );
+export default connect( mapStateToProps )( Exporter );
