+++ client/my-sites/exporter/notices.jsx
@@ -1,0 +1,70 @@
+/**
+ * External dependencies
+ */
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { localize } from 'i18n-calypso';
+
+/**
+ * Internal dependencies
+ */
+import Notice from 'components/notice';
+import NoticeAction from 'components/notice/notice-action';
+import support from 'lib/url/support';
+import { getExportingState } from 'state/site-settings/exporter/selectors';
+import { getSelectedSiteId } from 'state/ui/selectors';
+import { States } from 'state/site-settings/exporter/constants';
+
+/**
+ * Displays local notices for the Export tab of Site Settings
+ */
+class Notices extends Component {
+	render() {
+		const {
+			exportDidComplete,
+			exportDidFail,
+			exportDownloadURL,
+			translate,
+		} = this.props;
+
+		if ( exportDidComplete ) {
+			return (
+				<Notice
+					status="is-success"
+					showDismiss={ false }
+					text={ translate( `Your export was successful!
+						A download link has also been sent to your email.` ) }
+				>
+					<NoticeAction href={ exportDownloadURL }>
+						{ translate( 'Download' ) }
+					</NoticeAction>
+				</Notice>
+			);
+		}
+		if ( exportDidFail ) {
+			return (
+				<Notice
+					status="is-error"
+					showDismiss={ false }
+					text={ translate( `There was a problem preparing your
+						export file. Please check your connection and try
+						again, or contact support.` ) }
+				>
+					<NoticeAction href={ support.CALYPSO_CONTACT }>
+						{ translate( 'Get Help' ) }
+					</NoticeAction>
+				</Notice>
+			);
+		}
+
+		return null;
+	}
+}
+
+const mapStateToProps = state => ( {
+	exportDidComplete: getExportingState( state, getSelectedSiteId( state ) ) === States.COMPLETE,
+	exportDidFail: getExportingState( state, getSelectedSiteId( state ) ) === States.FAILED,
+	exportDownloadURL: state.siteSettings.exporter.downloadURL,
+} );
+
+export default connect( mapStateToProps )( localize( Notices ) );
